requirements.txt
python-telegram-bot 
aiohttp 
aiosocks
requests
pycryptodome
python-socketio[client]
websocket-client
